[PB-186] Tilpasset OppgaveQueriesTest til å bruke junit5
